@@ -16,6 +16,7 @@
     "terminal"
   ],
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "ava test",
     "eslint": "eslint **/*.js",
@@ -37,10 +38,10 @@
     "node": ">=6.4.0"
   },
   "devDependencies": {
-    "ava": "^0.25.0",
+    "ava": "^4.2.0",
     "dedent": "^0.7.0",
-    "eslint": "^4.19.1",
-    "nyc": "^11.7.3",
-    "rimraf": "^2.6.2"
+    "eslint": "^8.15.0",
+    "nyc": "^15.1.0",
+    "rimraf": "^3.0.2"
   }
 }
